error catch when editing customer

+++ BusinessLayer/CustomerService.py
@@ -36,8 +36,10 @@ class CustomerService:
 
 	def delete_customer(self, customer_email):
 		'''eyðir viðskiptavin með því að kalla á repóið'''
-		self.__order_repo.get_order(customer_email)
-		self.__customer_repo.remove_customer(customer_email)
+		customer = self.__order_repo.get_order(customer_email)
+		if customer == False:
+			self.__customer_repo.remove_customer(customer_email)
+		else:
 
 	def find_customer(self, customer_email):
 		return self.__customer_repo.find_customer_by_email(customer_email)
